+++ tests/test-deferred-symbol/test_qemu.c
@@ -28,7 +28,7 @@ static int test_deferred_single(udynlink_load_mode_t mode) {
 
     // Call test function (should return 1 because late_init is NULL)
     {
-        uint32_t *mod_base = (uint32_t *)UDYNLINK_LOT_BASE_ADDR;
+        uintptr_t *mod_base = (uintptr_t *)UDYNLINK_LOT_BASE_ADDR;
         *mod_base = mod.ram_base;
         udynlink_sym_t sym;
         if (udynlink_lookup_symbol(&mod, "test", &sym) == NULL) {
@@ -47,7 +47,7 @@ static int test_deferred_single(udynlink_load_mode_t mode) {
     }
 
     // Patch the symbol directly
-    if (udynlink_link_symbol(&mod, "late_init", (uint32_t)(uintptr_t)&real_late_init) != UDYNLINK_OK) {
+    if (udynlink_link_symbol(&mod, "late_init", (uintptr_t)&real_late_init) != UDYNLINK_OK) {
         printf("link_symbol failed\n");
         ok = 0;
         goto cleanup;
@@ -55,7 +55,7 @@ static int test_deferred_single(udynlink_load_mode_t mode) {
 
     // Call test function again (should return 2 and print "late_init called")
     {
-        uint32_t *mod_base = (uint32_t *)UDYNLINK_LOT_BASE_ADDR;
+        uintptr_t *mod_base = (uintptr_t *)UDYNLINK_LOT_BASE_ADDR;
         *mod_base = mod.ram_base;
         udynlink_sym_t sym;
         if (udynlink_lookup_symbol(&mod, "test", &sym) == NULL) {
